Revert "refactor: #714 내 토론 ViewModel 분리"
This reverts commit 14b31ac6de3507b8896960004268d82bff03a639.
