+++ android/TodokTodok/app/src/main/java/com/team/todoktodok/presentation/compose/discussion/DiscussionsActivity.kt
@@ -11,8 +11,6 @@ import androidx.activity.viewModels
 import com.team.todoktodok.App
 import com.team.todoktodok.presentation.compose.discussion.latest.vm.LatestDiscussionViewModel
 import com.team.todoktodok.presentation.compose.discussion.latest.vm.LatestDiscussionViewModelFactory
-import com.team.todoktodok.presentation.compose.discussion.my.vm.MyDiscussionViewModel
-import com.team.todoktodok.presentation.compose.discussion.my.vm.MyDiscussionViewModelFactory
 import com.team.todoktodok.presentation.compose.discussion.vm.DiscussionsViewModel
 import com.team.todoktodok.presentation.compose.discussion.vm.DiscussionsViewModelFactory
 import com.team.todoktodok.presentation.compose.theme.TodoktodokTheme
@@ -34,6 +32,7 @@ class DiscussionsActivity : ComponentActivity() {
         val repositoryModule = container.repositoryModule
         DiscussionsViewModelFactory(
             repositoryModule.discussionRepository,
+            repositoryModule.memberRepository,
             repositoryModule.notificationRepository,
             container.connectivityObserver,
         )
@@ -44,15 +43,6 @@ class DiscussionsActivity : ComponentActivity() {
         val repositoryModule = container.repositoryModule
         LatestDiscussionViewModelFactory(
             repositoryModule.discussionRepository,
-            container.connectivityObserver,
-        )
-    }
-
-    private val myDiscussionViewModel: MyDiscussionViewModel by viewModels {
-        val container = (application as App).container
-        val repositoryModule = container.repositoryModule
-        MyDiscussionViewModelFactory(
-            repositoryModule.memberRepository,
             container.connectivityObserver,
         )
     }
@@ -97,8 +87,7 @@ class DiscussionsActivity : ComponentActivity() {
             TodoktodokTheme {
                 DiscussionsScreen(
                     viewModel = viewModel,
-                    latestDiscussionViewModel = latestDiscussionViewModel,
-                    myDiscussionViewModel = myDiscussionViewModel,
+                    latestDiscussionViewModel,
                     exceptionMessageConverter = messageConverter,
                     onDiscussionClick = ::moveToDiscussionDetail,
                     onClickNotification = ::moveToNotification,
@@ -159,6 +148,11 @@ class DiscussionsActivity : ComponentActivity() {
     private fun moveToCreateDiscussion() {
         val intent = SelectBookActivity.Intent(this)
         startActivity(intent)
+    }
+
+    override fun onResume() {
+        super.onResume()
+        viewModel.loadMyDiscussions()
     }
 
     private fun handleNotificationDeepLink(intent: Intent) {
