+++ android/TodokTodok/app/src/main/java/com/team/todoktodok/presentation/compose/discussion/component/DiscussionTab.kt
@@ -34,7 +34,6 @@ import com.team.todoktodok.presentation.compose.discussion.model.Destination
 import com.team.todoktodok.presentation.compose.discussion.model.Destination.Companion.Destination
 import com.team.todoktodok.presentation.compose.discussion.model.DiscussionsUiState
 import com.team.todoktodok.presentation.compose.discussion.my.MyDiscussionsScreen
-import com.team.todoktodok.presentation.compose.discussion.my.vm.MyDiscussionViewModel
 import com.team.todoktodok.presentation.compose.discussion.search.SearchDiscussionScreen
 import com.team.todoktodok.presentation.compose.theme.Green1A
 import com.team.todoktodok.presentation.compose.theme.Pretendard
@@ -46,7 +45,6 @@ import kotlinx.coroutines.launch
 @Composable
 fun DiscussionTab(
     latestDiscussionViewModel: LatestDiscussionViewModel,
-    myDiscussionViewModel: MyDiscussionViewModel,
     messageConverter: ExceptionMessageConverter,
     uiState: DiscussionsUiState,
     pagerState: PagerState,
@@ -139,7 +137,6 @@ fun DiscussionTab(
                                     onClick = onClickDiscussion,
                                     modifier = Modifier.fillMaxSize(),
                                 )
-
                             AllDiscussionMode.SEARCH ->
                                 SearchDiscussionScreen(
                                     uiState = uiState.searchDiscussion,
@@ -150,9 +147,8 @@ fun DiscussionTab(
 
                     Destination.MY ->
                         MyDiscussionsScreen(
-                            viewModel = myDiscussionViewModel,
-                            messageConverter = messageConverter,
-                            onClickDiscussion = onClickDiscussion,
+                            uiState = uiState.myDiscussion,
+                            onClick = onClickDiscussion,
                             onClickHeader = onClickMyDiscussionHeader,
                             modifier = Modifier.fillMaxSize(),
                         )
