+++ android/TodokTodok/app/src/main/java/com/team/todoktodok/presentation/compose/discussion/vm/DiscussionsViewModelFactory.kt
@@ -4,10 +4,12 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.team.domain.ConnectivityObserver
 import com.team.domain.repository.DiscussionRepository
+import com.team.domain.repository.MemberRepository
 import com.team.domain.repository.NotificationRepository
 
 class DiscussionsViewModelFactory(
     private val discussionRepository: DiscussionRepository,
+    private val memberRepository: MemberRepository,
     private val notificationRepository: NotificationRepository,
     private val networkConnectivityObserver: ConnectivityObserver,
 ) : ViewModelProvider.Factory {
@@ -15,6 +17,7 @@ class DiscussionsViewModelFactory(
         if (modelClass.isAssignableFrom(DiscussionsViewModel::class.java)) {
             return DiscussionsViewModel(
                 discussionRepository,
+                memberRepository,
                 notificationRepository,
                 networkConnectivityObserver,
             ) as T
